requirements.txt
Flask==2.2.5
Flask-SQLAlchemy>=3.0.0
Flask-JWT-Extended==4.3.1
python-dotenv==0.20.0
psycopg[binary]
Werkzeug>=2.2.2
pytest==7.1.2
gunicorn
flask-swagger-ui
Flask-CORS
Flask-Migrate>=3.1.0
